requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-script
flask_restful
flask-cors
flask_jwt_extended
psycopg2-binary
gunicorn
pylint
flake8
pytest==4.3.1
pytest-cov
pytest-mock
pylint-flask
coveralls==1.7.0

# chat
flask-socketio
eventlet
passlib
firebase-admin

requests
